renamed NFC to AutoStart
