+++ Timer/res/layout/activity_stopwatch_autostart.xml
@@ -9,6 +9,6 @@
         android:layout_centerHorizontal="true"
         android:layout_centerVertical="true"
         android:text="@string/hello_world"
-        tools:context=".Stopwatch_NFC" />
+        tools:context=".StopwatchAutostartActivity" />
 
 </RelativeLayout>
