+++ Timer/res/values/strings.xml
@@ -16,5 +16,6 @@
     <string name="hello_world">Hello world!</string>
     <string name="menu_settings">Settings</string>
     <string name="title_activity_stopwatch__nfc">Stopwatch_NFC</string>
+    <string name="title_activity_stopwatch_autostart">StopwatchAutostartActivity</string>
 
 </resources>
